==> 01_precheck.Rmd <==
# Precheck

Make sure to use a full `polities.csv` file (if you re-run using the output the multiple imputation loop will run into bugs when it comes across NGAs with only one polity)

NB: Using shortened variables.csv file that only uses variables of interest to save computational time. If using full file, may need to renumber variables in Aggr.MI

**Don't rerun this section unless it's a new scrape**
  

# Checks Soc Complx data from exportdat.csv for error

The variable `polities` was used all over the code. MM changed the variable name to `polities_pre` for pre-checked, pre-tidied version in order to distinguish it with the one used for analysis. 

```{r}
polities_raw <- read_csv("input/polities.csv")
```

### Quick look at the raw data

```{r polities-quick-plot}
polities_raw %>% 
  transform(Complexity  = factor(Complexity, levels = c("Early", "Medium", "Late"))) %>% 
  ggplot(aes(x = NGA, ymin = Start, y = (Start + End) / 2, ymax = End, colour = Complexity)) +
  geom_errorbar(width = 0.3) +
  geom_text(aes(label = PolName, size = abs(End - Start)), colour = "Black", vjust = -0.1, show.legend = FALSE) + # angle = 90
  scale_size_continuous(range = c(0, 3)) +
  scale_colour_viridis_d() +
  coord_flip()

```

### delete duplicated entries 

`Dupl == y` is deleted **probably** so that an NGA will only have one polity at a time. 

```{r}


polities_pre <- polities_raw %>% 
  filter(Dupl == "n")
# quick-look again
polities_pre %>% 
    transform(Complexity  = factor(Complexity, levels = c("Early", "Medium", "Late"))) %>%  # just for plot sake
  ggplot(aes(x = NGA, ymin = Start, y = (Start + End) / 2, ymax = End, colour = Complexity)) +
  geom_errorbar(width = 0.3) +
  geom_text(aes(label = PolName, size = abs(End - Start)), colour = "Black", vjust = -0.1, show.legend = FALSE) + # angle = 90
  scale_size_continuous(range = c(0, 3)) +
  scale_colour_viridis_d() +
  coord_flip()

```

### Vars


`Vars` is a 225 row variables of complexity such as population, professional soldiers etc.

```{r}
Vars_df <- read_csv("input/variables.csv") %>% # originally as.matrix()
  mutate(Variable = paste(Subsection, Variable)) # Creating unique variable/section combinations
# Vars[,1] <- paste(Vars[,2],Vars[,1]) 
```


## dat filtering it into SCdat

`dat` seems to be a matrix that contain nearly 40k entries of changes in social complexity on variables such as ...wakaran! 
Not sure why matrix is used for `dat`. 
At first glance, `SCdat` seems to be a modified `dat` which is generated by filtering its `Variables` by `Vars`'s `Variables`. 
After that, it seems that scdat filters for the Polity that matches with `polities`'s PolID.
Is this correct??
The nrow() of the `dat` should shrink down to 25813 then 21025. Almost all `Comments` are NA. 34 rows are either "warning - Surrounding a simple value in uncertain brackets" or "The fact is uncertain" 

```{r}
# dat_old <- read.table('input/exportdat.csv', sep=",", header=TRUE, quote = "", colClasses = "character")
dat_tidy <- 
  read_csv("input/exportdat.csv", col_types = cols(`Value To` = col_character())) %>%  # without col_types, returns parsing failures on row 5342, 18249, 18750. maybe col_double is more suitable? DUNNO
  filter(Section %in% c("Social Complexity variables",  "Ritual variables", "Religion and Normative Ideology"))  %>%   # Section is set in !MI.R
  mutate(Variable = paste(Subsection, Variable))  %>% # ISSUE: str_c does not work: "a" + NA = NA. 
  filter(Variable %in% Vars_df$Variable) %T>% {print(dim(.))} %>%  
  filter(Polity %in% polities_pre$PolID) %T>% {print(dim(.))} %>% 
  select(-Section, -Subsection, -Comment)

# row.names(dat) <- NULL # at the first glance i dunno why we need this... we'll get back to this ISSUE
# nrow(SCdat) - sum(is.na(SCdat$Comment))
```


## Recode

Convert categorical values to numbers. Ignore warnings: they will be taken care off in the next step -- in errors

```{r}

dat_recoded <- dat_tidy

dat_recoded$`Value From` <- 
  dat_recoded$`Value From` %>% 
  recode(
    present = "1", 
    `inferred present` = "0.9",
    `present` ="1",   
    `inferred present` ="0.9",   
    `inferred absent` ="0.1",   
    `absent` ="0",   
    `present absent` = "0.5", # this seems to be dropped from the original Nature script, but had to be included. not sure how to code this though. tentatively putting 0.5.
    `none` ="0",
    `daily` ="6",   
    `weekly` ="5",
    `monthly` ="4",  
    `monthly/seasonal` = "3.5", # 3.5 maybe? this seems to be dropped from the original Nature script, but had to be included.
    `seasonally` ="3",  
    `yearly` ="2",   
    `once per generation` ="1",   
    `once in a generation` = "1", # this seems to be dropped from the original Nature script, but had to be included.
    `once in a lifetime` ="0",    
    `whole polity` ="3",      
    `majority` ="2",      
    `substantial minority` ="1",      
    `elites` ="0",          
    `inactive` ="1",
    `active` ="2",          
    `moralizing` ="3",          
    `monotheistic` ="1",     
    `polytheistic` ="2",                                                     
    `not applicable` = NA_character_,      
    `not applicalbe` =  NA_character_,  # this seems to be dropped from the original Nature script, but had to be included.
    `suspected unknown` =  NA_character_,      
    `suspected unknwon` = NA_character_, # this seems to be dropped from the original Nature script, but had to be included.
    `uncoded` = NA_character_, # this seems to be dropped from the original Nature script, but had to be included.
    `unknown` = NA_character_ # https://github.com/tidyverse/dplyr/issues/3202
  )
```

## inspect

test if there's still character

```{r}

# inspect old
dat_tidy$`Value From` %>% 
str_detect(letters) %>% sum() # 5096 rows contain letters.
dat_tidy$`Value To` %>% 
str_detect(letters) %>% sum() # NA

# inspect new
dat_recoded$`Value From` %>% 
str_detect(letters) %>% sum() # NA

dat_recoded$`Value To` %>% 
  str_detect(letters) %>% sum() # NA

```
none.


```{r}
SCdat_df <- dat_recoded %>%
  mutate(`Value From` = as.numeric(as.character(dat_recoded$`Value From`)),
         `Value To` = as.numeric(as.character(dat_recoded$`Value To`))) %>%  # Convert characters "100" to `100` by `as.character() %>% as.numeric()`. not sure if this is doing fine
  drop_na(`Value From`) %>% 
   mutate(`Date From` = `Date From` %>% str_replace("([0-9]+)BCE", "-\\1") %>% str_replace("([0-9]+)CE", "\\1") %>% as.integer() , 
         `Date To` = `Date To` %>% str_replace("([0-9]+)BCE", "-\\1") %>% str_replace("([0-9]+)CE", "\\1") %>% as.integer()
         ) # Change BCE to negative years and remove CE. 

# NAs introduced by coersion. is this okay? ISSUE
```

Let's take a quick look at the new `SCdat_df`.
```{r scdat-quicklook, fig.height = 20, fig.width = 20}
SCdat_df %>% 
  filter(NGA == "Kansai") %>% 
  ggplot(aes(x = `Date From`, y = `Value From`, colour = `Date From`)) +
  geom_point() +
  # geom_line() +
  scale_colour_viridis_c() +
  facet_wrap(~Variable, scale = "free_y")
```
